requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.8.3
python-dotenv==1.0.0
pyodbc==5.0.1
pymssql==2.3.1
requests==2.31.0
pydapper==0.8.0
